FormatStringConverter: Refactor
This is rather a large refactoring of FormatStringConverter.

Get rid of FormatStringResult. It grew out of a simple
Optional<std::string> to something more painful. Rather than store up
all the information to hand back to the check, let's just make the check
create a FormatStringConverter itself and then call it to apply the
fixes when we're ready. This gives the check the ability to make its own
fixes first (required for printf and strprintf.)

The trace fixer has been renamed slightly and the column number for the
fix is now for the call rather than the format string because that's
easier to get hold of from the check.

The various FormatArgOffset variables are now actually correct -
previously they were badly named as the value was actually the index of
the first argument _after_ the format string. However,
FormatStringConverter::ArgsOffset is the offset of the arguments for
formatting because it's easier to implement HandlePrintfSpecifier that
way.