TraceConvertCheck: Convert from TRACE to TRACE.F
When I originally wrote this converter I expected to run it at the same
time as converting the implementation of the operator() method of the
trace class from printf-style to fmt-style format strings.

It turns out that rolling the fix out to a large amount of code is
easier if the fix is idempotent. The easiest way to do this is to
convert the operator() method call to a call to the existing F() member
function. Once all callers have been converted (and we can prove this by
removing the operator() method) then the code can hopefully be switched
back to using operator() this time expecting a fmt-style format string
with a simple search and replace operation. If that doesn't work, then
writing a fix-it check to do so ought to be straightforward.