[clang-tidy] readability-redundant-string-cstr for smart pointer #576705
Fixes: https://github.com/llvm/llvm-project/issues/56705

The readability-redundant-string-cstr check incorrectly replaces calls
to c_str() that involve an overloaded operator->.

Running `clang-tidy --checks="-*,readability-redundant-string-cstr"` on:

 #include <string>
 #include <vector>

 void it(std::vector<std::string>::const_iterator i)
 {
   std::string tmp;
   tmp = i->c_str();
 }

yields:

 /home/mac/git/llvm-project/build/../bug.cpp:7:9: warning: redundant call to 'c_str' [readability-redundant-string-cstr]
   tmp = i->c_str();
         ^~~~~~~~~~
         *i->

which means that the code is "fixed" to incorrectly say:
   tmp = *i->;
rather than the expected:
   tmp = *i;

This appears to be due to the overloaded `operator->` meaning that
RedundantStringCStrCheck.cpp#L53 ends up with Text containing "i->"
rather than just the expected "i".

Add test case for this and fix it in a somewhat nasty manner in the
absence of something better.