[clang-tidy] Add modernize-trace-format check
Use the same FormatStringConverter as used the printf to fmt::print
converter to convert calls to a Trace class operator()(const char *,...)
method to use std::format format strings and call a method named
FMT_STYLE on the same object. It is expected that once the conversion is
complete a simple search and replace can be used to switch these back to
operator() calls in one go. The conversion can't be performed whilst
leaving the function name the same since the check wouldn't then be
idempotent.