[clang-tidy] TRACE convert from printf format to {fmt} format
Use the same FormatStringConverter as used the printf to fmt::print
converter to convert calls to a Trace class operator()(const char *)
method to use {fmt} format strings. Note that the call itself is not
changed, so it is assumed that it will be updated to expect such strings
at the same time.