Implemented this in the lexer directly., This required making a few methods in Preprocessor public
This works except with nested f-literals in f-literals. This does not come out right as the token order gets jumbled because the inner MacroInfo object gets pushed first, instead of being pushed as part of evaluation of the outer macro object.

Another problem with this simple approach as that it doesn't handle concatenation of literals correctly, i.e. it doesn't collect all argument expressions after the last of a sequence of string literals, a subset of which may be of (the same) prefix kind.