Tweak macros passed on command line to avoid confusing lit on Windows
On at least Windows, the `%s` in the macros passed on the command line
for the tests are being replaced with the filename for the test. Let's
try to avoid that by removing the `%` prefix from all the test macros.
This makes them consistent with the <inttypes.h> macros too.