Entire test program runs.
Replaced the five new token kinds with a token flag IsFLiteral.

Handle :> when ] not expected as : >

Pass the f of f-literals in StringLiteralParser::init

Implemented the formation of a std::format call from all the collected expressions and literals with all literals before all expressions in the Parser::ParseStringLiteralExpression function. Then it recurses to ParseCastExpression.

Removed comma error as we can't do this in the lexer, or template argument lists won't work.